% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bds_estimation_new.R
\name{c1}
\alias{c1}
\title{Calculate c1 parameter for birth-death-sampling model}
\usage{
c1(lambda, mu, psi)
}
\arguments{
\item{lambda}{Speciation rate}

\item{mu}{Extinction rate}

\item{psi}{Sampling rate}
}
\value{
Value of c1
}
\description{
Calculate c1 parameter for birth-death-sampling model
}
\keyword{internal}
