% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bds_estimation_new.R
\name{c2}
\alias{c2}
\title{Calculate c2 parameter for birth-death-sampling model}
\usage{
c2(lambda, mu, psi)
}
\arguments{
\item{lambda}{Speciation rate}

\item{mu}{Extinction rate}

\item{psi}{Sampling rate}
}
\value{
Value of c2
}
\description{
Calculate c2 parameter for birth-death-sampling model
}
\keyword{internal}
