% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_new.R
\name{create_params_template}
\alias{create_params_template}
\title{Creates a data frame for saasi}
\usage{
create_params_template(
  states,
  lambda = 0.5,
  mu = 0.01,
  psi = 0.5,
  prior = NULL
)
}
\arguments{
\item{states}{Vector of state names or number of states}

\item{lambda}{Speciation rate per unit time (single value or vector per state)}

\item{mu}{Extinction rate per unit time (single value or vector per state)}

\item{psi}{Sampling rate per unit time (single value or vector per state)}

\item{prior}{Prior probabilities (default: uniform; or vector per state), should sum up to 1.}
}
\value{
A data frame for saasi
}
\description{
Creates a data frame for saasi
}
\examples{
params <- create_params_template(
  states = c("A", "B"),
  lambda = 0.5,
  mu = 0.01,
  psi = c(0.1, 0.9)  
)
}
