% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generator_segment.R, R/utilityfunctions.R
\name{create_psi_stepwise}
\alias{create_psi_stepwise}
\title{Create time-dependent psi parameters using simple data frame approach}
\usage{
create_psi_stepwise(time_points, psi_matrix)

create_psi_stepwise(time_points, psi_matrix)
}
\arguments{
\item{time_points}{The time step}

\item{psi_matrix}{sampling rate matrix}
}
\value{
Data frame ready to use with saasi

A sampling rate
}
\description{
Create time-dependent psi parameters using simple data frame approach

Sampling rate
}
