% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_metadata}
\alias{demo_metadata}
\title{SAASI demonstration: tree metadata}
\format{
A \code{data.frame} with 13 rows (one per tip) and 2 columns
\describe{
\item{\code{node}}{Node labels for each tip}
\item{\code{state}}{Character state for each tip}
}
}
\usage{
demo_metadata
}
\description{
A \code{data.frame} linking the tip nodes to their state
}
\examples{
data(demo_metadata)
demo_metadata

}
\keyword{datasets}
