% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_pars}
\alias{demo_pars}
\title{SAASI demonstration: birth-death-sampling process parameters}
\format{
A data frame with 3 rows (one per state) and 5 columns:
\describe{
\item{\code{state}}{The character label for each state}
\item{\code{root_prior}}{Prior probability distribution over states for the root node}
\item{\code{lambda}}{Numeric speciation rate for each state}
\item{\code{mu}}{Numeric extinction rate for each state}
\item{\code{psi}}{Numeric Sampling rate for each state}
}
}
\usage{
demo_pars
}
\description{
A \code{data.frame} specifying the parameters of the birth-death-sampling process that is used to
simulate the demonstration tree. These parameters control diversification
and sampling rates for each state.
}
\examples{
data(demo_pars)
demo_pars
}
\keyword{datasets}
