% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_tree}
\alias{demo_tree}
\title{SAASI demonstration: timed phylogenetic tree}
\format{
An object of class \code{phylo}
}
\usage{
demo_tree
}
\description{
A simulated birth-death-sampling tree with 13 tips.
The tip states are located separately in the \link{demo_metadata} file.
}
\examples{
demo_tree
}
\keyword{datasets}
