% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_tree_prepared}
\alias{demo_tree_prepared}
\title{SAASI demonstration: pre-prepared tree}
\format{
An object of class \code{phylo} with \code{tip.state} attribute
}
\usage{
demo_tree_prepared
}
\description{
A simulated birth-death-sampling tree with 13 tips and 3 states.
The tree has pre-prepared to be compatible with \code{saasi}.
}
\examples{
data(demo_tree_prepared)
demo_tree_prepared$tip.state
}
\keyword{datasets}
