% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebola_tree}
\alias{ebola_tree}
\title{Ebola tree example}
\format{
A \code{phylo} tree object
}
\source{
Downloaded from Nextstrain
}
\usage{
ebola_tree
}
\description{
An ebola tree with 1493 tips and 3 states used as example.
}
\examples{
data(ebola_tree)
ebola_tree$root.edge
}
\keyword{datasets}
