% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bds_estimation_new.R
\name{estimate_a}
\alias{estimate_a}
\title{Estimate net diversification rate from lineage-through-time data}
\usage{
estimate_a(phy, trim = c(0.1, 0.5))
}
\arguments{
\item{phy}{A phylo object with branch lengths}

\item{trim}{Numeric vector of length 2 specifying quantiles for trimming.
Default is c(0.10, 0.50), using the middle 40\% of node times.}
}
\value{
Estimated net diversification rate (a = lambda - mu - psi)
}
\description{
Estimates the parameter 'a' (lambda - mu - psi) by fitting a linear model
to the log-transformed lineage counts over time (LTT), using a trimmed portion
of the internal node times.
}
\keyword{internal}
