% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_new.R
\name{estimate_transition_rates}
\alias{estimate_transition_rates}
\title{Estimates transition rates using ace or simmap.
Default using ace, if fails use simmap instead}
\usage{
estimate_transition_rates(tree, model = "ER", custom_q = NULL, method = "ace")
}
\arguments{
\item{tree}{A phylo object with tip.state attribute}

\item{model}{Transition model: "ER" (equal rates), "SYM" (symmetric),
"ARD" (all rates different), or "custom" (default: "ER")}

\item{custom_q}{Custom Q matrix (only used if model = "custom")}

\item{method}{Which method to use.}
}
\value{
A transition rate matrix (Q matrix) ready for SAASI
}
\description{
Estimates transition rates using ace or simmap.
Default using ace, if fails use simmap instead
}
\examples{
# Symmetric model
data(ebola_tree)
q_matrix <- estimate_transition_rates(ebola_tree, model = "SYM")
}
