% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{extract_ace_q}
\alias{extract_ace_q}
\title{Construct a rate matrix q using the output of ace}
\usage{
extract_ace_q(aceoutput)
}
\arguments{
\item{aceoutput}{the output of ace}
}
\value{
A rate matrix q in the form saasi needs
}
\description{
This function is here because ace returns rates and an index matrix, not a matrix of rates.
}
