% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bds_estimation_new.R
\name{fit_bd_fixed_mu}
\alias{fit_bd_fixed_mu}
\title{Fit birth-death-sampling model with fixed extinction rate}
\usage{
fit_bd_fixed_mu(
  phy,
  mu,
  n_starts = 10,
  psi_max = 7,
  r0_min = 1,
  r0_max = 5,
  infectious_period_min = NULL,
  infectious_period_max = NULL
)
}
\arguments{
\item{phy}{A phylo object with branch lengths}

\item{mu}{Fixed extinction rate}

\item{n_starts}{Number of starting points}

\item{psi_max}{Maximum sampling rate}

\item{r0_min}{Minimum R0}

\item{r0_max}{Maximum R0}

\item{infectious_period_min}{Minimum infectious period (1/(mu+psi))}

\item{infectious_period_max}{Maximum infectious period (1/(mu+psi))}
}
\value{
List containing estimated parameters and diagnostics
}
\description{
Estimates lambda (speication rate) and psi (sampling rate) using maximum likelihood
with a fixed mu (extinction rate), subject to some epidemiological constraints.
}
\keyword{internal}
