% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generator_seg.R,
%   R/tree_generator_segment.R
\name{get_psi_at_sim_time}
\alias{get_psi_at_sim_time}
\title{Get psi values at specific simulation time using forward time (FIXED)}
\usage{
get_psi_at_sim_time(sim_time, psi_df, max_t)

get_psi_at_sim_time(sim_time, psi_df, max_t)
}
\arguments{
\item{sim_time}{Current simulation time (forward time from 0)}

\item{psi_df}{Data frame with time-dependent psi values}

\item{max_t}{Maximum simulation time}
}
\value{
Vector of psi values for each state at the given simulation time

Vector of psi values for each state at the given simulation time
}
\description{
Get psi values at specific simulation time using forward time (FIXED)

Get psi values at specific simulation time using step function
}
