% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generator_segment.R, R/utilityfunctions.R
\name{get_psi_at_time_stepwise}
\alias{get_psi_at_time_stepwise}
\title{Get psi values at specific time using step function (constant within intervals)}
\usage{
get_psi_at_time_stepwise(time_point, psi_df)

get_psi_at_time_stepwise(time_point, psi_df)
}
\arguments{
\item{time_point}{Numeric value representing the time point}

\item{psi_df}{Data frame with time-dependent psi values.
First column: time, remaining columns: psi values for each state}

\item{time_points}{The time step}

\item{psi_matrix}{sampling rate matrix}
}
\value{
Vector of psi values for each state at the given time point

A sampling rate
}
\description{
Get psi values at specific time using step function (constant within intervals)

Helper function to get psi at time for stepwise intervals
MODIFIED to work correctly with the new stepwise structure
}
