% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bds_estimation_new.R
\name{log_ge}
\alias{log_ge}
\title{Calculate log-likelihood under the general birth-death-sampling model}
\usage{
log_ge(phy, lambda, mu, psi)
}
\arguments{
\item{phy}{A phylo object}

\item{lambda}{Speication rate}

\item{mu}{Extinction rate}

\item{psi}{Sampling rate}
}
\value{
Log-likelihood value
}
\description{
Calculate log-likelihood under the general birth-death-sampling model
}
\keyword{internal}
