% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bds_estimation_new.R
\name{log_q}
\alias{log_q}
\title{Calculate log of q function for birth-death-sampling model}
\usage{
log_q(t, lambda, mu, psi)
}
\arguments{
\item{t}{Time values}

\item{lambda}{Speication rate}

\item{mu}{Extinction rate}

\item{psi}{Sampling rate}
}
\value{
Log of q function values
}
\description{
Calculate log of q function for birth-death-sampling model
}
\keyword{internal}
