% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bds_estimation_new.R
\name{logsumexp}
\alias{logsumexp}
\title{Numerically stable log-sum-exp function}
\usage{
logsumexp(a, b, c)
}
\arguments{
\item{a}{First log value}

\item{b}{Second log value}

\item{c}{Third log value}
}
\value{
log(exp(a) + exp(b) + exp(c))
}
\description{
Numerically stable log-sum-exp function
}
\keyword{internal}
