% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generator_seg.R,
%   R/tree_generator_segment.R
\name{make.tree.bisse_modified_time_psi}
\alias{make.tree.bisse_modified_time_psi}
\title{FIXED: Tree simulation with time-dependent psi (original function name)}
\usage{
make.tree.bisse_modified_time_psi(
  pars,
  k,
  max.taxa = Inf,
  max.t = Inf,
  x0,
  psi_df = NULL
)

make.tree.bisse_modified_time_psi(
  pars,
  k,
  max.taxa = Inf,
  max.t = Inf,
  x0,
  psi_df = NULL
)
}
\arguments{
\item{pars}{Parameter matrix (lambda, mu, psi, q_ij...)}

\item{k}{Number of states}

\item{max.taxa}{Maximum number of taxa}

\item{max.t}{Maximum simulation time}

\item{x0}{Root state}

\item{psi_df}{Optional data frame with time-dependent psi values}
}
\description{
FIXED: Tree simulation with time-dependent psi (original function name)

MODIFIED: Tree simulation with time-dependent psi
}
