% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_mu_generator.R
\name{mle_lm}
\alias{mle_lm}
\title{Estimate speciation/extinction rates for a tree}
\usage{
mle_lm(phy, lambda = 2, mu = 0.5, psi = 0.5, method = "L-BFGS-B")
}
\arguments{
\item{phy}{A \code{phylo} phylogenetic tree (\code{ape} format).}

\item{lambda}{An initial "guess" for speciation, used in subsequent formulae.}

\item{mu}{An initial "guess" for extinction, used in subsequent formulae.}

\item{psi}{Sampling rate.}

\item{method}{See \code{method} parameter in \link{mle}.}
}
\value{
A two-element vector containing speciation and extinction values
estimated from \code{phy}.
}
\description{
This is done by method of maximum likelihood. Mostly implemented by
\href{https://github.com/yexuansong}{@yexuansong}, from methods described in
\href{https://doi.org/10.1093/molbev/msr217}{Stadler et al. (2012)}.
}
