% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_estimate.R
\name{mle_lm_including_psi}
\alias{mle_lm_including_psi}
\title{Estimate speciation/extinction rates for a tree}
\usage{
mle_lm_including_psi(
  phy,
  lambda,
  mu,
  psi,
  method = "L-BFGS-B",
  lower = c(0.001, 0.001, 0.001),
  upper = c(2, 1, 1)
)
}
\arguments{
\item{phy}{A \code{phylo} phylogenetic tree (\code{ape} format).}

\item{lambda}{An initial "guess" for speciation, used in subsequent formulae.}

\item{mu}{An initial "guess" for extinction, used in subsequent formulae.}

\item{psi}{Sampling rate.}

\item{method}{See \code{method} parameter in \link{mle}.}

\item{lower}{A three-element vector containing the lower bound of the speciation,
extinction, and sampling rate values (greater than 0), the default is c(0.001,0.001,0.001).}

\item{upper}{A three-element vector containing the upper bound of the speciation,
extinction, and sampling rate values.}
}
\value{
A fitted mle object containing speciation, extinction, and sampling rate estimates.
}
\description{
This is done with a maximum likelihood method, implemented mainly by
\href{https://github.com/yexuansong}{@yexuansong}, from methods described in
\href{https://doi.org/10.1093/molbev/msr217}{Stadler et al. (2012)}.
}
