% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_new.R
\name{plot_saasi}
\alias{plot_saasi}
\title{Plot a phylogenetic tree with SAASI tip annotations}
\usage{
plot_saasi(
  tree,
  saasi_result,
  colours = NULL,
  tip_cex = 0.5,
  node_cex = 0.2,
  save_file = NULL,
  width = 3000,
  height = 3000,
  res = 300
)
}
\arguments{
\item{tree}{An object of class \code{phylo} with \code{tip.state} attribute.}

\item{saasi_result}{Output of \code{saasi()} containing internal node annotations.}

\item{colours}{Optional character vector specifying the node colour for each state. By default, colours are automatically
generated based on the number of states.}

\item{tip_cex}{Size of the tips. Default value is 0.5.}

\item{node_cex}{Size of node pie charts. Default value is 0.2.}

\item{save_file}{Optional string specifying the file path where the plot will be saved.
By default the plot is displayed on the current device.}

\item{width}{Width of the plot in pixels. Only relevant if saving to file. Default value is 3000 pixels.}

\item{height}{Height of the plot in pixels. Only relevant if saving to file. Default value is 3000 pixels.}

\item{res}{Plot resolution in dpi. Only relevant if saving to file. Default value is 300 dpi.}
}
\value{
If a save file is specified, the plot will be saved as directed.
If not, then the plot will be displayed on the users device.
The function returns the output of \code{plot.phylo()}, a list containing plot specifications.
}
\description{
\code{plot_saasi()} plots a phylogenetic tree annoted with \code{\link[=saasi]{saasi()}} output. Tips are annotated with their states.
Internal nodes are annotated with a pie chart showing the marginal distribution over states for each node.
The \code{\link[ape:plot.phylo]{ape::plot.phylo()}} function is used for plotting.
}
\examples{
# Run SAASI
saasi_res <- saasi(demo_tree_prepared, demo_Q, demo_pars)

# Plot the results using the default settings
plot_saasi(demo_tree_prepared, saasi_res)

\dontrun{
# Use custom colours and save the result to tree.png
plot_saasi(tree, saasi_result,
           colours = c("orange", "darkblue", "pink"),
           node_cex = 0.3,
           save_file = "tree.png")
}

}
