% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{q_adjust}
\alias{q_adjust}
\title{Transition rate adjustment}
\usage{
q_adjust(qij_matrix, state, sampling_diff)
}
\arguments{
\item{qij_matrix}{The matrix representation of the transition rates, can be
obtained using \code{extract_ace_q} function.}

\item{state}{The state that has higher or lower sampling rate relative to
the other states. If \code{state} is not a character but an index (numeric value), pick
the index relate to the index of the matrix.}

\item{sampling_diff}{Sampling difference the state of interest and other states.
This should be a single positive value indicates the sampling difference (ratio).
If 0<sampling_diff<1, this implies that
the state of interest samples less than other states. If
sampling_diff > 1, this implies that the state of interest samples more than other
states.}
}
\value{
A new transition rate matrix that is adjusted by the sampling differences.
}
\description{
This function adjust the estimated transition rates using \code{ace} based on
different sampling fractions.
}
