% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saasi.R
\name{saasi}
\alias{saasi}
\title{Sampling Aware Ancestral State Inference}
\usage{
saasi(phy, Q, pars, sensitivity_test = FALSE)
}
\arguments{
\item{phy}{A \code{phylo} object containing the phylogenetic tree. The tree must be
rooted, binary, and contain \code{tip.state}.}

\item{Q}{A numeric transition rate matrix \exp{(n \times n)} where n is the number
of states. Row and column names represent states. Off-diagonal
elements are transition rates; diagonal elements are set such that rows sum to zero.}

\item{pars}{A data frame with the other parameters used in the ancestral state reconstruction algorithm.
Must have the following column names: state, prior, lambda, mu, and psi.
The prior values refer to the baseline probabilities of the states (used at the root of the tree).}
}
\value{
A data frame with state probabilities for each internal node in \code{phy}.
}
\description{
Reconstructs ancestral states for internal nodes of a phylogenetic tree while
accounting for heterogeneous sampling rates across states or locations.
}
\examples{
head(demo_pars) #contains state, lambda, mu, psi for each state
result <- saasi(phy = demo_tree_prepared, 
                Q = demo_Q,
                pars = demo_pars) 
head(result) 
}
