% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bds_tree.R
\name{sim_bds_tree}
\alias{sim_bds_tree}
\title{Simulate a phylogenetic tree from a birth-death-sampling process}
\usage{
sim_bds_tree(
  params_df,
  q_matrix,
  x0,
  max_taxa = 100,
  max_t = 100,
  include_extinct = FALSE,
  min_tip = 1
)
}
\arguments{
\item{params_df}{A \code{data.frame} specifying the parameters from which the tree will be simulated. Must contain
\itemize{
\item \code{lambda}: vector of speciation rates
\item \code{mu}: vector of extinction rates
\item \code{psi}: vector of sampling rates
}

The number of rows should be equal to the number of states.}

\item{q_matrix}{State transition rate matrix.}

\item{x0}{The root state.}

\item{max_taxa}{Maximum number of nodes allowed in the initial simulated tree. Default value is 100.}

\item{max_t}{Maximum depth allowed in the returned tree. Default value is 100.}

\item{include_extinct}{Boolean. If \code{TRUE}, extinct taxa are included in the returned tree. Default value is \code{FALSE}.}

\item{min_tip}{Minimum number of tips required in the post-processed tree.
If the processed tree has fewer tips, the simulation is repeated until this
condition is satisfied. Default value is 1.}
}
\value{
An object of class \code{phylo} that is compatible with \code{saasi()}.
}
\description{
\code{sim_bds_tree()} simulates a phylogenetic tree with state annotations for all internal nodes and tips.
A birth-death-sampling process is used to generate the tree topology and branch lengths;
this will include speciation, sampling, and mutation events.
A continuous-time Markov process is used to assign states to all nodes.
The resulting tree will be compatible with \code{\link[=saasi]{saasi()}}.
The tree is post-processed to remove tips at the present and ensure a minimum number of tips.
}
\examples{
# Simulate a tree from the demo model with root state set to 1
demo_pars
demo_Q
sim_bds_tree(demo_pars, demo_Q, 1)
}
