% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generator_seg.R,
%   R/tree_generator_segment.R
\name{sim_bds_tree_time_psi}
\alias{sim_bds_tree_time_psi}
\title{FIXED: Simulate tree with time-dependent psi (original function name)}
\usage{
sim_bds_tree_time_psi(
  params_df,
  q_matrix,
  x0,
  max_taxa = 100,
  max_t = 100,
  include_extinct = FALSE,
  psi_df = NULL
)

sim_bds_tree_time_psi(
  params_df,
  q_matrix,
  x0,
  max_taxa = 100,
  max_t = 100,
  include_extinct = FALSE,
  psi_df = NULL
)
}
\arguments{
\item{params_df}{Data frame with parameters. Can have time-dependent psi.}

\item{q_matrix}{Transition rate matrix}

\item{x0}{Root state}

\item{max_taxa}{Maximum number of taxa}

\item{max_t}{Maximum simulation time}

\item{include_extinct}{Include extinct lineages}

\item{psi_df}{Optional: Data frame with time-dependent psi values.}
}
\value{
A phylo object with time-dependent sampling history
}
\description{
FIXED: Simulate tree with time-dependent psi (original function name)

You'll need your existing me.to.ape.bisse function here
(Copy it from your original tree_generator.R file)
MODIFIED: Simulate a birth/death/sampling tree with time-dependent psi
}
