% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tip_states.R
\name{add_tip_states}
\alias{add_tip_states}
\title{add tip states to phylogenetic tree}
\usage{
add_tip_states(tree, tip_data)
}
\arguments{
\item{tree}{A phylo object}

\item{tip_data}{A named vector where names = tip labels, values = states, or
a df with 2 columns: column 1 = tip labels, column 2 = states}
}
\value{
An object of class \code{phylo} with annotated tip states.
}
\description{
\code{add_tip_states()} annotates the tips of a \code{phylo} object with state values.
}
\details{
A \code{tip.states} attribute is added to \code{tree} based on the \code{tip_data} provided by the user.
This function is used within the \code{prepare_tree_for_saasi()} function.
}
\examples{
demo_tree
head(demo_metadata)

add_tip_states(demo_tree, demo_metadata)

}
\seealso{
\code{\link[=drop_tips_by_state]{drop_tips_by_state()}} to remove specific traits or \code{\link[=prepare_tree_for_saasi]{prepare_tree_for_saasi()}} for reformatting \code{phylo} objects to be \code{saasi}-compatible.
}
