% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_preparation_new.R
\name{add_tip_states}
\alias{add_tip_states}
\title{add tip states to phylogenetic tree}
\usage{
add_tip_states(tree, tip_data)
}
\arguments{
\item{tree}{A phylo object}

\item{tip_data}{A named vector where names = tip labels, values = states, or
a df with 2 columns: column 1 = tip labels, column 2 = states}
}
\value{
A phylo object with tip.state added
}
\description{
add tip states to phylogenetic tree
}
