% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asrproject.R
\name{asr}
\alias{asr}
\title{Test package.}
\usage{
asr(phy, params_file, q_matrix_file)
}
\arguments{
\item{phy}{TODO required attrs}

\item{params_file}{TODO}

\item{q_matrix_file}{TODO}
}
\value{
TODO phylo
}
\description{
Test package.
}
