% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generator_segment.R, R/utilityfunctions.R
\name{create_psi_stepwise}
\alias{create_psi_stepwise}
\title{Create time-dependent psi parameters using simple data frame approach}
\usage{
create_psi_stepwise(time_points, psi_matrix)

create_psi_stepwise(time_points, psi_matrix)
}
\arguments{
\item{time_points}{Vector of time points where psi changes (boundary points)}

\item{psi_matrix}{Matrix where each row corresponds to psi values for each interval
Should have (length(time_points) + 1) rows for stepwise intervals}
}
\value{
Data frame ready to use with saasi

Data frame ready to use with saasi
}
\description{
Create time-dependent psi parameters using simple data frame approach
}
\details{
For stepwise psi functions:
\itemize{
\item n time points create n+1 intervals
\item psi_matrix should have n+1 rows (one for each interval)
}

Example:
time_points = c(5, 10) creates intervals [0,5), [5,10), [10,∞)
psi_matrix should have 3 rows for these 3 intervals
}
