% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_metadata}
\alias{demo_metadata}
\title{Metadata for demonstration tree}
\format{
A data frame with 13 rows (one per tip) and 2 columns
\describe{
\item{node}{node name for each tip}
\item{state}{Character state for each tip (1, 2, or 3)}
}
}
\usage{
demo_metadata
}
\description{
A data frame containing the node name and tip states
}
\examples{
data(demo_metadata)
demo_metadata
demo_metadata$states

}
\keyword{datasets}
