% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_Q}
\alias{demo_Q}
\title{Transition rate matrix for demonstration tree}
\format{
A 3x3 numeric matrix
}
\usage{
demo_Q
}
\description{
A 3x3 transition rate matrix (Q matrix) used to simulate the demonstration
tree. This matrix defines the instantaneous rates of transition between
character states.
}
\examples{
data(demo_Q)
demo_Q
# Check that rows sum to 0
rowSums(demo_Q)
}
\keyword{datasets}
