% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_tree_prepared}
\alias{demo_tree_prepared}
\title{Simulated demonstration tree example}
\format{
A \code{phylo} tree object with the following components
}
\usage{
demo_tree_prepared
}
\description{
A simulated birth-death-sampling tree with 13 tips and 3 states.
}
\examples{
data(demo_tree_prepared)
demo_tree_prepared$tip.state
}
\keyword{datasets}
