% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_preparation_new.R
\name{drop_tips_by_state}
\alias{drop_tips_by_state}
\title{Drop tips by state}
\usage{
drop_tips_by_state(tree, drop_values)
}
\arguments{
\item{tree}{A phylo object with tip.state}

\item{drop_values}{Character vector of states to remove.
Include NA to drop tips with NA states.
Example: c("Not Collected", NA)}
}
\value{
A phylo object with the specified tips removed and
tip.state correctly updated.
}
\description{
Removes tips that matches any of the specified values.
}
