% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bds_estimation_new.R
\name{estimate_bds_parameters}
\alias{estimate_bds_parameters}
\title{Estimate rate parameters of a birth-death-sampling process}
\usage{
estimate_bds_parameters(
  phy,
  mu,
  trim = c(0.1, 0.5),
  n_starts = 100,
  force_two_step = FALSE,
  psi_max = 7,
  r0_min = 1,
  r0_max = 5,
  infectious_period_min = NULL,
  infectious_period_max = NULL
)
}
\arguments{
\item{phy}{An object of class \code{phylo} with branch lengths. Must be rooted and binary.}

\item{mu}{The extinction rate. This must be non-negative.}

\item{trim}{A numeric vector of length 2 specifying the quantiles for removing
node times in the lineages-through-time regression (see Details). Default value is \code{c(0.10, 0.50)}.}

\item{n_starts}{The number of starting points for multi-start optimization. Default value is \code{100}.}

\item{force_two_step}{Logical. If \code{TRUE}, uses two-step estimates even if they
violate constraints (see Details). Default value is \code{FALSE}.}

\item{psi_max}{Maximum sampling rate per unit time. Must be strictly positive. Default value is \code{7}.}

\item{r0_min}{Minimum basic reproductive number (R0 = lambda/(mu+psi)). Must be strictly positive. Default 1.}

\item{r0_max}{Maximum basic reproductive number. Must be strictly positive. Default 5.}

\item{infectious_period_min}{Minimum infectious period (1/(mu+psi)). Must be strictly positive. Default NULL.}

\item{infectious_period_max}{Maximum infectious period (1/(mu+psi)). Must be strictly positive. Default NULL.}
}
\value{
A list with the following attributes:
\itemize{
\item \code{lambda}: Estimated speciation rate
\item \code{psi}: Estimated sampling rate
\item \code{r0}: Estimated basic reproductive number (\eqn{\lambda/(\mu+\psi)})
\item \code{infectious_period}: Estimated infectious period (\eqn{1/(\mu+\psi)})
\item \code{net_diversification}: Estimated net diversification rate (\eqn{\lambda-\mu-\psi})
}
}
\description{
\code{estimate_bds_parameters()} estimates the speciation rate \eqn{\lambda} and sampling rate \eqn{\psi} of a birth-death-sampling process given a phylogenetic tree and death rate \eqn{\mu}.
The function uses user-specified epidemiological constraints on the sampling rate, basic reproduction number, and the duration of the infectious period to improve estimability.
}
\details{
Estimation is carried out by first estimating the reproduction number \eqn{\frac{\lambda}{\mu+\psi}} and the net diversification rate \eqn{\lambda-\mu-\psi},
then solving for the unknown \eqn{\lambda} and \eqn{\psi}. The net diversification rate is estimated by
performing a lineages-through-time (LTT) regression on node times that are within the quantiles specified in the \code{trim} input
and the reproduction number is estimated by first computing maximum likelihood
estimates (MLEs) of \eqn{\lambda} and \eqn{\psi} by numerically optimizing a likelihood and then compute the reproduction number.
The stability of the MLEs is assessed by checking if the estimate of \eqn{R_0} is within 0.02 of 1.

If the estimates of \eqn{\lambda} and \eqn{\psi} do not satisfy all of the constraints specified by the \code{psi_max}, \code{r0_min}, \code{r0_max}, \code{infectious_period_min},
and \code{infectious_period_max} parameters, then the MLEs of \eqn{\lambda} and \eqn{\psi} are returned, unless \code{force_two_step=TRUE}
in which case the original estimates of \eqn{\lambda} and \eqn{\psi} are returned regardless of
whether or not the constraints are satisfied.
If LTT regression fails then the MLEs of \eqn{\lambda} and \eqn{\psi} will be returned.

The MLEs are obtained numerically using the \code{L-BFGS-S} algorithm (see \link[stats::optim()]{optim} for implementation details) \code{n_starts} times from randomly chosen starting points
and selecting the best maximizer of all attempts. Due to the user-imposed constraints, not all randomly generated starting
points are valid, so \code{100*n_starts} attempts are made to generate valid starting points. This may result in fewer than \code{n_starts} optimizations
actually being performed.

If the two-step method produces estimates that violate constraints or is
numerically unstable, the function falls back to the MLE estimates from step 1.
}
\examples{
data(ebola_tree)

# Use the following information about the 2013 Ebola outbreak to obtain estimates
# - Average removal rate of 5
# - Infectious periods range from 20 to 40 days
# - An upper bound on the sampling rate of 15
# - Plausible values for the basic reproduction number are between 1.5 and 3
BDS_fit <- estimate_bds_parameters(ebola_tree, mu = 5,
                                               psi_max = 15,
                                               infectious_period_min = 20/365,
                                               infectious_period_max = 40/365,
                                               r0_min = 1.5,
                                               r0_max = 3)

}
\seealso{
\code{\link[=estimate_transition_rates]{estimate_transition_rates()}}
}
