% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_new.R
\name{estimate_transition_rates}
\alias{estimate_transition_rates}
\title{Estimate the transition rate matrix of the discrete trait process}
\usage{
estimate_transition_rates(tree, matrix_structure = "ER", method = "ace")
}
\arguments{
\item{tree}{A phylo object with a tip.state attribute assigning traits to all tips.}

\item{matrix_structure}{The form of the transition rate matrix. Can either be a numeric matrix or one of the following strings: \code{"ER"}, \code{"SYM"}, and \code{"ARD"} (see details). The default value is \code{"ER"}.}

\item{method}{The method used to perform the estimation. Possible values are \code{"ace"} or \code{"fitMk"}.}
}
\value{
A transition rate matrix that has been fit to the observed phylogeny. This matrix will be compatible with other \code{saasi} functions.
}
\description{
\code{estimate_transition_rates()} computes a maximum likelihood estimate of the transition rate matrix \eqn{Q} governing the discrete trait process.
The discrete traits evolve according to a Markov process with transition rate matrix \eqn{Q} such that element \eqn{q_{ij}} is the instantaneous rate of transitioning from state \eqn{i} to state \eqn{j}.
}
\details{
The \code{model} parameters controls the structure of \eqn{Q}.
\code{"ER"} specifies equal rates for all transitions, \eqn{q_{ij}=q} for all traits \eqn{i,j}.
\code{"SYM"} specifies a symmetric transition rate matrix, \eqn{q_{ij}=q_{ji}} for all traits \eqn{i,j}.
\code{"ARD"} places no structural constraints and allows all traits to be different, \eqn{q_{ij}\ne q_{kl}} for all traits \eqn{i,j,k,l}.
If another grouping of transition rates is desired this can be input as a numeric matrix such that each entry has a matrix with a positive integer and all entries with the same value will share a rate.

The \code{method} parameter allows the user to select a preferred method between \code{ace} and \code{fitMk}, however if the preferred method fails the other will be used instead.
If \code{ace} is selected, the maximum likelihood estimator is computed using the \code{\link[ape:ace]{ape::ace()}} function.
If \code{fitMk} is selected, the \code{\link[phytools:fitMk]{phytools::fitMk()}} function is used to fit \eqn{Q} to the provided phylogenetic tree.
}
\examples{
# Load a timed phylogenetic tree for Ebola
data(ebola_tree)

# Use the simmap function to estimate the rate transition matrix
# - Impose a symmetric form on the matrix
Q <- estimate_transition_rates(ebola_tree, matrix_structure = "SYM")

# Use ace to estimate the rate transition matrix
# - Impose a structure such that transitions to or from Guinea happen at a different rate than Liberia or Sierra Leone
struct <- matrix(c(0, 1, 1, 1, 0, 2, 1, 2, 0), nrow=3, ncol=3)
Q <- estimate_transition_rates(ebola_tree, matrix_structure = struct)
}
