% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generator.R
\name{get_rand_phy}
\alias{get_rand_phy}
\title{TODO}
\usage{
get_rand_phy(seed, max_taxa, params_file, q_matrix_file, plot = FALSE)
}
\arguments{
\item{seed}{TODO}

\item{max_taxa}{TODO}

\item{params_file}{TODO}

\item{q_matrix_file}{TODO}

\item{plot}{TODO}
}
\value{
TODO
}
\description{
two_sampled_states_phy <-
get_rand_phy(3, 20, "two_sampled_states_params.csv",
"two_sampled_states_q.csv", plot = TRUE)
}
