% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_mu_generator.R
\name{mle_lm}
\alias{mle_lm}
\title{Estimate speciation/extinction rates for a tree}
\usage{
mle_lm(
  phy,
  lambda,
  mu,
  psi,
  method = "L-BFGS-B",
  lower = c(0.001, 0.001),
  upper
)
}
\arguments{
\item{phy}{A \code{phylo} phylogenetic tree (\code{ape} format).}

\item{lambda}{An initial "guess" for speciation, used in subsequent formulae.}

\item{mu}{An initial "guess" for extinction, used in subsequent formulae.}

\item{psi}{Sampling rate.}

\item{method}{See \code{method} parameter in \link{mle}.}

\item{lower}{A two-element vector containing the lower bound of the speciation
and extinction values (geater than 0), the default is c(0.001,0.001).}

\item{upper}{A two-element vector containing the upper bound of the speciation
and extinction values, should be similar to the sampling rate, still need to find
a better justification for the upcoming version.}
}
\value{
A two-element vector containing speciation and extinction values
estimated from \code{phy}.
}
\description{
This is done by method of maximum likelihood. Mostly implemented by
\href{https://github.com/yexuansong}{@yexuansong}, from methods described in
\href{https://doi.org/10.1093/molbev/msr217}{Stadler et al. (2012)}.
}
