% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_preparation_new.R
\name{prepare_tree_for_saasi}
\alias{prepare_tree_for_saasi}
\title{Prepare phylogenetic tree for saasi}
\usage{
prepare_tree_for_saasi(
  tree,
  tip_data = NULL,
  resolve_polytomies = TRUE,
  fix_branches = TRUE,
  min_branch_length = 1e-05,
  drop_states = character(0)
)
}
\arguments{
\item{tree}{A phylo object}

\item{tip_data}{A named vector where names = tip labels, values = states, or
a df with 2 columns: column 1 = tip labels, column 2 = states.
Optional if tree already has tip states}

\item{resolve_polytomies}{Resolve polytomies with multi2di}

\item{fix_branches}{Fix zero/negative branch lengths}

\item{min_branch_length}{Replace value for zero or negative branch lengths (default: 1e-5)}

\item{drop_states}{States to be removed from the tree. NA is always included.
Add additional values to drop states. Example:c("Not Collected").}
}
\value{
A prepared phylo object compatible to SAASI.
}
\description{
Prepare phylogenetic tree for saasi
}
