% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tree_for_saasi.R
\name{prepare_tree_for_saasi}
\alias{prepare_tree_for_saasi}
\title{Prepare phylogenetic tree for SAASI}
\usage{
prepare_tree_for_saasi(
  tree,
  tip_data = NULL,
  resolve_polytomies = TRUE,
  fix_branches = TRUE,
  min_branch_length = 1e-05,
  drop_states = character(0)
)
}
\arguments{
\item{tree}{A tree of class \code{phylo}.}

\item{tip_data}{A named vector where names = tip labels, values = states, or
a dataframe with 2 columns: column 1 = tip labels, column 2 = states.
If tree already contains tip states they should be named \code{tip.state}.}

\item{resolve_polytomies}{Resolve polytomies with \link[ape:multi2di]{ape::multi2di}.}

\item{fix_branches}{Resolve zero/negative branch lengths.}

\item{min_branch_length}{Replace value for zero or negative branch lengths (default: 1e-5)}

\item{drop_states}{States to be removed from the tree. NA is always included.
Add additional values to drop states. Example:c("Not Collected").}
}
\value{
A prepared \code{phylo} object to be used as input for \link{saasi} and/or \link{plot_saasi}.
}
\description{
Prepare phylogenetic tree for SAASI
}
\examples{
data(ebola_tree)
# this tree already has tip states 
ebola_tree$tip.state
ebola_tree_prep <- prepare_tree_for_saasi(ebola_tree)

}
