% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saasi.R
\name{saasi}
\alias{saasi}
\title{Sampling Aware Ancestral State Inference}
\usage{
saasi(phy, params_df, q_matrix)
}
\arguments{
\item{phy}{A \code{phylo} phylogenetic tree. The tree needs to be a rooted binary tree. Must contain
\code{tip.state}. \code{tip.state} can be names of the states, or numeric values (1:n).}

\item{params_df}{Data frame containing non-q parameters used in ancestral
state reconstruction algorithm. Must have the following column names:
\code{state}, \code{prior}, \code{lambda}, \code{mu}, and \code{psi}. The \code{prior} values refer to the baseline
probabilities of the states (used at the root of the tree).

Example:\tabular{lllll}{
   \strong{state} \tab \strong{prior} \tab \strong{lambda} \tab \strong{mu} \tab \strong{psi} \cr
   1 \tab 1/3 \tab 3 \tab 0.02 \tab 1 \cr
   2 \tab 1/3 \tab 3 \tab 0.02 \tab 1 \cr
   3 \tab 1/3 \tab 3 \tab 0.02 \tab 1 \cr
}}

\item{q_matrix}{Numeric q matrix used in ancestral state reconstruction
algorithm. Row and column indices or names represent states.}
}
\value{
A data frame listing the state probabilities of every node in \code{phy}.
}
\description{
Get the internal node state probabilities of a tree with defined leaf states.
}
