% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saasi.R
\name{saasi}
\alias{saasi}
\title{Sampling Aware Ancestral State Inference}
\usage{
saasi(phy, q_matrix, lambda, mu, psi, prior = NULL)
}
\arguments{
\item{phy}{The output of \link{prepare_tree_for_saasi}, i.e., a rooted, timed, phylogentic tree of class \code{phylo} with states assigned to each tip.}

\item{q_matrix}{A named matrix. The \eqn{n\times n} stochastic rate matrix used in ancestral state reconstruction. Row and column names should correspond to states.}

\item{lambda}{Speciation rate per unit time. Numeric value (if equal for all states) or a vector of length equal to the number of states containing a value for each state.}

\item{mu}{Extinction rate per unit time. Numeric value (if equal for all states) or a vector of length equal to the number of states containing a value for each state.}

\item{psi}{Sampling rate per unit time. Numeric value (if equal for all states) or a vector of length equal to the number of states containing a value for each state.}

\item{prior}{A vector of length equal to the number of states containing prior probabilities for each state. The vector must sum to 1. The \code{prior} values refer to the baseline
probabilities of the states (used at the root of the tree).}
}
\value{
A data frame of size (number of internal nodes by number of states) containing state probabilities of every node in \code{phy}. Each column represents a state. The row names correspond to the node IDs.
}
\description{
Compute internal node state probabilities of a tree with defined tip states.
}
\examples{
data(ebola_tree) 
# note: ebola_tree was already prepared using prepare_tree_for_saasi
unique(ebola_tree$tip.state)
# ebola_tree has 3 states
ebola_q <- matrix(c(-3,   1,   2,
                    0.5, -2,  1.5,
                    1,   0.5, -1.5),
 nrow = 3,
 byrow = TRUE
)
rownames(ebola_q) = unique(ebola_tree$tip.state)
colnames(ebola_q) = unique(ebola_tree$tip.state)

phi = rep(0.5,3)
lambda = rep(0.6,3)
mu = rep(0.2,3)
psi = rep(0.2,3)

ebola_saasi <- saasi(ebola_tree, ebola_q, phi, lambda, mu, psi)
head(ebola_saasi)
}
