% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saasi.R
\name{saasi}
\alias{saasi}
\title{Sampling Aware Ancestral State Inference}
\usage{
saasi(phy, q_matrix, lambda, mu, psi, prior = NULL)
}
\arguments{
\item{phy}{The output of \code{prepare_tree_for_saasi}, i.e., a rooted, timed, phylogentic tree of class \code{phylo} with states assigned to each tip.}

\item{q_matrix}{A named matrix. The \eqn{n\times n} stochastic rate matrix used in ancestral state reconstruction. Row and column column names in the same order as \code{states}.}

\item{lambda}{Speciation rate per unit time. Numeric value (if equal for all states) or a vector of length equal to the number of states containing a value for each state.}

\item{mu}{Extinction rate per unit time. Numeric value (if equal for all states) or a vector of length equal to the number of states containing a value for each state.}

\item{psi}{Sampling rate per unit time. Numeric value (if equal for all states) or a vector of length equal to the number of states containing a value for each state.}

\item{prior}{A vector of length equal to the number of states containing prior probabilities for each state. The vector must sum to 1. The \code{prior} values refer to the baseline
probabilities of the states (used at the root of the tree).}

\item{states}{Vector of state names.}
}
\value{
A data frame listing the state probabilities of every node in \code{phy}. The row names correspond to the node IDs.
}
\description{
Compute internal node state probabilities of a tree with defined leaf states.
}
