% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saasi.R
\name{saasi}
\alias{saasi}
\title{Sampling Aware Ancestral State Inference}
\usage{
saasi(phy, params_df, q_matrix, plot = FALSE, cex = 1)
}
\arguments{
\item{phy}{A \code{phylo} phylogenetic tree (ape format). Must contain
\code{tip.state}.}

\item{params_df}{Data frame containing non-q parameters used in ancestral
state reconstruction algorithm. Must have the following column names:
\code{state}, \code{freq}, \code{lambda}, \code{mu}, and \code{psi}. The state values should be a
1-based sequence of natural numbers: \code{1}, \code{2}, ..., \emph{\code{n}}.

Example:\tabular{lllll}{
   \strong{state} \tab \strong{freq} \tab \strong{lambda} \tab \strong{mu} \tab \strong{psi} \cr
   1 \tab 0.2 \tab 3 \tab 0.02 \tab 1 \cr
   2 \tab 0.3 \tab 3 \tab 0.02 \tab 1 \cr
   3 \tab 0.5 \tab 3 \tab 0.02 \tab 1 \cr
}}

\item{q_matrix}{Numeric q matrix used in ancestral state reconstruction
algorithm. Row and column indices represent states.}

\item{plot}{If \code{TRUE}, a plot of \code{phy} with overlaid state labels will be
rendered.}

\item{cex}{See \code{cex} parameter in \link{nodelabels}.}
}
\value{
A data frame listing the state probabilities of every node in \code{phy}.
}
\description{
Get the internal node state probabilities of a tree with defined leaf states.
}
