% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generator.R
\name{sim_bds_tree}
\alias{sim_bds_tree}
\title{Simulate a birth/death/sampling tree}
\usage{
sim_bds_tree(
  params_df,
  q_matrix,
  x0,
  max_taxa = 100,
  max_t = 100,
  include_extinct = FALSE
)
}
\arguments{
\item{q_matrix}{A named matrix. The \eqn{n\times n} stochastic rate matrix used in ancestral state reconstruction. Row and column column names in the same order as \code{states}.}

\item{x0}{Natural number used as root state in the returned tree. Must be a
state declared in \code{params_df}.}

\item{max_taxa}{Maximum number of nodes allowed in the returned tree.}

\item{max_t}{Maximum depth allowed in the returned tree.}

\item{include_extinct}{Boolean declaring whether extinct taxa are included in
the returned tree.}
}
\value{
A \code{phylo} phylogenetic tree (\code{ape} format).
}
\description{
This tree can be passed to \link{saasi}, and will include speciation,
extinction, sampling, and mutation events.
}
