% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generator.R
\name{sim_bds_tree}
\alias{sim_bds_tree}
\title{Simulate a birth/death/sampling tree}
\usage{
sim_bds_tree(
  params_df,
  q_matrix,
  x0,
  max_taxa = 100,
  max_t = 100,
  include_extinct = FALSE
)
}
\arguments{
\item{params_df}{Data frame containing non-q parameters used in ancestral
state reconstruction algorithm. Must have the following column names:
\code{state}, \code{freq}, \code{lambda}, \code{mu}, and \code{psi}. The state values should be a
1-based sequence of natural numbers: \code{1}, \code{2}, ..., \emph{\code{n}}.

Example:\tabular{lllll}{
   \strong{state} \tab \strong{freq} \tab \strong{lambda} \tab \strong{mu} \tab \strong{psi} \cr
   1 \tab 0.2 \tab 3 \tab 0.02 \tab 1 \cr
   2 \tab 0.3 \tab 3 \tab 0.02 \tab 1 \cr
   3 \tab 0.5 \tab 3 \tab 0.02 \tab 1 \cr
}}

\item{q_matrix}{Numeric q matrix used in ancestral state reconstruction
algorithm. Row and column indices represent states.}

\item{x0}{Natural number used as root state in the returned tree. Must be a
state declared in \code{params_df}.}

\item{max_taxa}{Maximum number of nodes allowed in the returned tree.}

\item{max_t}{Maximum depth allowed in the returned tree.}

\item{include_extinct}{Boolean declaring whether extinct taxa are included in
the returned tree.}
}
\value{
A \code{phylo} phylogenetic tree (\code{ape} format).
}
\description{
This tree can be passed to \link{saasi}, and will include speciation,
extinction, sampling, and mutation events.
}
